* Fixed compilationerrors in MSVC 7.1
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13886 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/emit/object/emit.cpp
@@ -921,7 +921,7 @@ void csEmitMeshObject::ReplaceAge(int time, const csColor& color, float alpha,
   csEmitAge *p = aging;
   while(p && (p->time != time))
   {
-    if(fabs(p->time-time) < SMALL_EPSILON)
+    if(fabs((float)p->time-time) < SMALL_EPSILON)
       break;  /// almost the same, assume the diff is due to float inaccuracy
     p=p->next;
   }
